requirements..txt
kivy
kivymd
opencv-python
pytesseract
Pillow
numpy

